refactor: . Move tests/Status* to tests/Statuses/
