+++ tests/Statuses/StatusSettingsReport.test.StatusSettingsReport_should_tabulate_StatusSettings.approved.md
